change something

+++ CoreDataLearn/ZLPhotoLib/Classes/ZLPhotoBrowser/Views/ZLPhotoPickerBrowserPhotoScrollView.m
@@ -85,7 +85,7 @@
     }
 }
 
-- (void)actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex{
+- (void)actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(long)buttonIndex{
     if (buttonIndex == 0){
         if([UIImagePickerController isSourceTypeAvailable: UIImagePickerControllerSourceTypeSavedPhotosAlbum]) {
             ALAssetsLibrary *assetsLibrary = [[ALAssetsLibrary alloc]init];
